requirements.txt
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.0.0
flask==2.3.3
scipy==1.10.1
scikit-image==0.21.0